Move caching To CachedTrait
todo getItemsByPage() must override by caching
