+++ src/Paginator.php
@@ -13,8 +13,6 @@ use ArrayIterator;
 use Countable;
 use IteratorAggregate;
 use Traversable;
-use Zend\Cache\Storage\IteratorInterface as CacheIterator;
-use Zend\Cache\Storage\StorageInterface as CacheStorage;
 use Zend\Filter\FilterInterface;
 use Zend\Paginator\Adapter\AdapterInterface;
 use Zend\Paginator\ScrollingStyle\ScrollingStyleInterface;
@@ -66,20 +64,6 @@ class Paginator implements Countable, IteratorAggregate
      * @var ScrollingStylePluginManager
      */
     protected static $scrollingStyles = null;
-
-    /**
-     * Cache object
-     *
-     * @var CacheStorage
-     */
-    protected static $cache;
-
-    /**
-     * Enable or disable the cache by Zend\Paginator\Paginator instance
-     *
-     * @var bool
-     */
-    protected $cacheEnabled = true;
 
     /**
      * Adapter
@@ -206,16 +190,6 @@ class Paginator implements Countable, IteratorAggregate
     }
 
     /**
-     * Sets a cache object
-     *
-     * @param CacheStorage $cache
-     */
-    public static function setCache(CacheStorage $cache)
-    {
-        static::$cache = $cache;
-    }
-
-    /**
      * Sets the default scrolling style.
      *
      * @param  string $scrollingStyle
@@ -297,18 +271,6 @@ class Paginator implements Countable, IteratorAggregate
     }
 
     /**
-     * Enables/Disables the cache for this instance
-     *
-     * @param bool $enable
-     * @return Paginator
-     */
-    public function setCacheEnabled($enable)
-    {
-        $this->cacheEnabled = (bool) $enable;
-        return $this;
-    }
-
-    /**
      * Returns the number of pages.
      *
      * @return int
@@ -330,34 +292,6 @@ class Paginator implements Countable, IteratorAggregate
     public function getTotalItemCount()
     {
         return count($this->getAdapter());
-    }
-
-    /**
-     * Clear the page item cache.
-     *
-     * @param int $pageNumber
-     * @return Paginator
-     */
-    public function clearPageItemCache($pageNumber = null)
-    {
-        if (!$this->cacheEnabled()) {
-            return $this;
-        }
-
-        if (null === $pageNumber) {
-            $prefixLength  = strlen(self::CACHE_TAG_PREFIX);
-            $cacheIterator = static::$cache->getIterator();
-            $cacheIterator->setMode(CacheIterator::CURRENT_AS_KEY);
-            foreach ($cacheIterator as $key) {
-                if (substr($key, 0, $prefixLength) == self::CACHE_TAG_PREFIX) {
-                    static::$cache->removeItem($this->_getCacheId((int)substr($key, $prefixLength)));
-                }
-            }
-        } else {
-            $cleanId = $this->_getCacheId($pageNumber);
-            static::$cache->removeItem($cleanId);
-        }
-        return $this;
     }
 
     /**
@@ -567,13 +501,6 @@ class Paginator implements Countable, IteratorAggregate
     {
         $pageNumber = $this->normalizePageNumber($pageNumber);
 
-        if ($this->cacheEnabled()) {
-            $data = static::$cache->getItem($this->_getCacheId($pageNumber));
-            if ($data) {
-                return $data;
-            }
-        }
-
         $offset = ($pageNumber - 1) * $this->getItemCountPerPage();
 
         $items = $this->adapter->getItems($offset, $this->getItemCountPerPage());
@@ -586,11 +513,6 @@ class Paginator implements Countable, IteratorAggregate
 
         if (!$items instanceof Traversable) {
             $items = new ArrayIterator($items);
-        }
-
-        if ($this->cacheEnabled()) {
-            $cacheId = $this->_getCacheId($pageNumber);
-            static::$cache->setItem($cacheId, $items);
         }
 
         return $items;
@@ -671,29 +593,6 @@ class Paginator implements Countable, IteratorAggregate
     }
 
     /**
-     * Returns the page item cache.
-     *
-     * @return array
-     */
-    public function getPageItemCache()
-    {
-        $data = [];
-        if ($this->cacheEnabled()) {
-            $prefixLength  = strlen(self::CACHE_TAG_PREFIX);
-            $cacheIterator = static::$cache->getIterator();
-            $cacheIterator->setMode(CacheIterator::CURRENT_AS_VALUE);
-            foreach ($cacheIterator as $key => $value) {
-                if (substr($key, 0, $prefixLength) == self::CACHE_TAG_PREFIX) {
-                    $data[(int) substr($key, $prefixLength)] = $value;
-                }
-            }
-        }
-        return $data;
-    }
-
-
-
-    /**
      * Brings the item number in range of the page.
      *
      * @param  int $itemNumber
@@ -735,51 +634,6 @@ class Paginator implements Countable, IteratorAggregate
         }
 
         return $pageNumber;
-    }
-
-    /**
-     * Tells if there is an active cache object
-     * and if the cache has not been disabled
-     *
-     * @return bool
-     */
-    protected function cacheEnabled()
-    {
-        return ((static::$cache !== null) && $this->cacheEnabled);
-    }
-
-    /**
-     * Makes an Id for the cache
-     * Depends on the adapter object and the page number
-     *
-     * Used to store item in cache from that Paginator instance
-     *  and that current page
-     *
-     * @param int $page
-     * @return string
-     */
-    protected function _getCacheId($page = null)
-    {
-        if ($page === null) {
-            $page = $this->getCurrentPageNumber();
-        }
-        return self::CACHE_TAG_PREFIX . $page . '_' . $this->_getCacheInternalId();
-    }
-
-    /**
-     * Get the internal cache id
-     * Depends on the adapter and the item count per page
-     *
-     * Used to tag that unique Paginator instance in cache
-     *
-     * @return string
-     */
-    protected function _getCacheInternalId()
-    {
-        return md5(serialize([
-            spl_object_hash($this->getAdapter()),
-            $this->getItemCountPerPage()
-        ]));
     }
 
     /**
